requirements.txt
numpy 
pillow 
opencv-python 
shapely 
trimesh 
